requirements.txt
Adafruit-Blinka==7.2.1
adafruit-blinka-displayio==0.10.2
adafruit-circuitpython-busdevice==5.1.8
adafruit-circuitpython-displayio-ssd1306==1.5.3
adafruit-circuitpython-framebuf==1.4.9
adafruit-circuitpython-requests==1.12.12
adafruit-circuitpython-ssd1306==2.12.4
adafruit-circuitpython-typing==1.5.0
Adafruit-PlatformDetect==3.22.1
Adafruit-PureIO==1.1.9
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
colorzero==2.0
crccheck==1.3.0
exceptiongroup==1.1.0
flake8==6.0.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gpiozero==1.6.2
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.4
influxdb-client==1.36.0
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
minio==7.1.13
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pyflakes==3.0.1
pyftdi==0.54.0
pyserial==3.5
pyTelegramBotAPI==4.10.0
pytest==7.2.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-telegram-bot==20.1
pyusb==1.2.1
reactivex==4.0.4
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
types-python-dateutil==2.8.19.6
typing_extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
